@@ -10,24 +10,30 @@
         }
     ],
     "require": {
-        "php": ">=5.3",
-        "coisa/exceptions": "~1.0",
-        "psr/container": "^1.0"
+        "php": "^7.4 || ^8.0",
+        "coisa/exceptions": "^1.0",
+        "psr/container": "^1.0 || ^2.0"
     },
     "require-dev": {
-        "coisa/php-cs-fixer": "^1.0",
+        "coisa/php-cs-fixer": "^2.1",
         "doctrine/annotations": "^1.2",
-        "jakub-onderka/php-console-highlighter": "^0.3.2 || ^0.4",
-        "jakub-onderka/php-parallel-lint": "^1.0",
-        "phpunit/phpunit": "^4.8 || ^7.0",
-        "squizlabs/php_codesniffer": "^2.9 || ^3.0"
+        "php-parallel-lint/php-console-highlighter": "^1.0",
+        "php-parallel-lint/php-parallel-lint": "^1.3",
+        "phpspec/prophecy-phpunit": "^2.0",
+        "phpunit/phpunit": "^9.5",
+        "sensiolabs/security-checker": "^6.0",
+        "squizlabs/php_codesniffer": "^3.6"
     },
     "config": {
+        "platform": {
+            "php": "7.4"
+        },
+        "prefer-stable": true,
         "sort-packages": true
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "1.x-dev"
+            "dev-2.x": "2.0"
         }
     },
     "autoload": {
@@ -46,12 +52,14 @@
     "minimum-stability": "stable",
     "scripts": {
         "check": [
+            "@lint",
             "@test",
             "@sniff",
             "@cs-check"
         ],
         "cs-check": "php-cs-fixer fix --dry-run --diff",
         "cs-fix": "php-cs-fixer fix",
+        "lint": "parallel-lint --exclude vendor .",
         "sniff": "phpcs",
         "test": "phpunit --color=always"
     }
